Revert "Feature #1001 - Starname Registration Flow"
