+++ packages/bierzo-wallet/src/routes/addresses/components/StarnamesNotExists.tsx
@@ -3,10 +3,9 @@ import { useTheme } from "@material-ui/styles";
 import { Block, Typography } from "medulas-react-components";
 import React from "react";
 
-import { history } from "../..";
 import { RpcEndpointType } from "../../../communication/rpcEndpoint";
-import { REGISTER_NAME_ROUTE, REGISTER_STARNAME_ROUTE } from "../../paths";
-import { NoStarnameHeader } from "../../register/components/StarnameForm";
+import { REGISTER_PERSONALIZED_ADDRESS_ROUTE } from "../../paths";
+import { NoStarnameHeader } from "../../registerStarName";
 
 interface StarnamesNotExistsProps {
   readonly onRegisterStarname: () => void;
@@ -30,7 +29,7 @@ export function GetYourAddressWithExtension({
       </Typography>
       <Block marginTop={3} />
       <Typography
-        id={REGISTER_STARNAME_ROUTE}
+        id={REGISTER_PERSONALIZED_ADDRESS_ROUTE}
         variant="subtitle1"
         color="primary"
         weight="semibold"
@@ -39,19 +38,6 @@ export function GetYourAddressWithExtension({
         onClick={onRegisterStarname}
       >
         Register Now
-      </Typography>
-      {/* TODO remove this Typography when /register-name accessible from starname list */}
-      <Typography
-        variant="subtitle1"
-        color="primary"
-        weight="semibold"
-        inline
-        link
-        onClick={() => {
-          history.push(REGISTER_NAME_ROUTE);
-        }}
-      >
-        Register your name
       </Typography>
     </Block>
   );
@@ -65,7 +51,7 @@ export function GetYourAddressWithLedger(): JSX.Element {
       <Typography variant="body1" weight="light">
         You can not register
       </Typography>
-      <Typography id={REGISTER_STARNAME_ROUTE} variant="body1" color="primary" weight="light">
+      <Typography id={REGISTER_PERSONALIZED_ADDRESS_ROUTE} variant="body1" color="primary" weight="light">
         starname
       </Typography>
       <Block textAlign="center">
