+++ packages/bierzo-wallet/src/routes/addresses/index.stories.tsx
@@ -8,7 +8,7 @@ import DecoratedStorybook, { bierzoRoot } from "../../utils/storybook";
 import {
   REGISTER_USERNAME_REGISTRATION_STORY_PATH,
   REGISTER_USERNAME_STORY_PATH,
-} from "../register/index.stories";
+} from "../registerName/index.stories";
 import { BwUsernameWithChainName } from ".";
 import AddressesTab from "./components/AddressesTab";
 import Iovnames from "./components/Iovnames";
