+++ packages/bierzo-wallet/src/routes/addresses/index.tsx
@@ -9,7 +9,7 @@ import { RootState } from "../../store/reducers";
 import { getRpcEndpointType } from "../../store/rpcendpoint/selectors";
 import { BwUsername } from "../../store/usernames";
 import { getChainAddressPairWithNames } from "../../utils/tokens";
-import { REGISTER_IOVNAME_ROUTE, REGISTER_STARNAME_ROUTE } from "../paths";
+import { REGISTER_PERSONALIZED_ADDRESS_ROUTE, REGISTER_STARNAME } from "../paths";
 import AddressesTab from "./components/AddressesTab";
 
 export interface BwUsernameWithChainName extends BwUsername {
@@ -17,11 +17,11 @@ export interface BwUsernameWithChainName extends BwUsername {
 }
 
 function onRegisterUsername(): void {
-  history.push(REGISTER_IOVNAME_ROUTE);
+  history.push(REGISTER_PERSONALIZED_ADDRESS_ROUTE);
 }
 
 function onRegisterStarname(): void {
-  history.push(REGISTER_STARNAME_ROUTE);
+  history.push(REGISTER_STARNAME);
 }
 
 const Addresses = (): JSX.Element => {
