+++ packages/bierzo-wallet/src/routes/addresses/test/travelToReceivePayment.ts
@@ -3,7 +3,7 @@ import { Page } from "puppeteer";
 import { ADDRESSES_TEXT } from "../../../components/Header/components/LinksMenu";
 import { whenOnNavigatedToE2eRoute } from "../../../utils/test/navigation";
 import { ADDRESSES_ROUTE } from "../../paths";
-import { REGISTER_IOVNAME_VIEW_ID } from "../../register/components/IovnameForm";
+import { REGISTER_USERNAME_VIEW_ID } from "../../registerName/components";
 import { yourAddresses, yourIovnames } from "../components/AddressesTab";
 import { iovnamesViewId } from "../components/Iovnames";
 import { yourBlockchainAddressesId } from "../components/UserAddresses";
@@ -29,5 +29,5 @@ export async function travelToStarnamesTabE2E(page: Page): Promise<void> {
 export async function openFirstStarnameForEditingE2E(page: Page, username: string): Promise<void> {
   const [firstStarnameEditLink] = await page.$x(`//h6[contains(., 'Edit')]`);
   await firstStarnameEditLink.click();
-  await page.waitForSelector(`#${REGISTER_IOVNAME_VIEW_ID}`);
+  await page.waitForSelector(`#${REGISTER_USERNAME_VIEW_ID}`);
 }
