+++ packages/bierzo-wallet/src/routes/balance/index.stories.tsx
@@ -9,7 +9,7 @@ import DecoratedStorybook, { bierzoRoot } from "../../utils/storybook";
 import {
   REGISTER_USERNAME_REGISTRATION_STORY_PATH,
   REGISTER_USERNAME_STORY_PATH,
-} from "../register/index.stories";
+} from "../registerName/index.stories";
 import Layout from "./components/index";
 
 export const BALANCE_STORY_PATH = `${bierzoRoot}/Balance`;
