+++ packages/bierzo-wallet/src/routes/index.tsx
@@ -12,15 +12,15 @@ import {
   LOGIN_ROUTE,
   PAYMENT_ROUTE,
   POLICY_ROUTE,
-  REGISTER_IOVNAME_ROUTE,
-  REGISTER_NAME_ROUTE,
-  REGISTER_STARNAME_ROUTE,
+  REGISTER_PERSONALIZED_ADDRESS_ROUTE,
+  REGISTER_STARNAME,
   TERMS_ROUTE,
   TRANSACTIONS_ROUTE,
 } from "./paths";
 import Payment from "./payment";
 import Policy from "./policy";
-import Register, { ToRegister } from "./register";
+import RegisterUsername from "./registerName";
+import RegisterStarname from "./registerStarName";
 import Terms from "./terms";
 import Transactions from "./transactions";
 
@@ -36,13 +36,8 @@ const Routes = (): JSX.Element => (
         <Route exact path={ADDRESSES_ROUTE} component={Addresses} />
         <Route exact path={TRANSACTIONS_ROUTE} component={Transactions} />
         <Route exact path={BALANCE_ROUTE} component={Balance} />
-        <Route exact path={REGISTER_IOVNAME_ROUTE} render={() => <Register entity={ToRegister.Iovname} />} />
-        <Route
-          exact
-          path={REGISTER_STARNAME_ROUTE}
-          render={() => <Register entity={ToRegister.Starname} />}
-        />
-        <Route exact path={REGISTER_NAME_ROUTE} render={() => <Register entity={ToRegister.Name} />} />
+        <Route exact path={REGISTER_PERSONALIZED_ADDRESS_ROUTE} component={RegisterUsername} />
+        <Route exact path={REGISTER_STARNAME} component={RegisterStarname} />
         <Route exact path={TERMS_ROUTE} component={Terms} />
         <Route exact path={POLICY_ROUTE} component={Policy} />
       </RequireLogin>
