+++ packages/bierzo-wallet/src/routes/paths.ts
@@ -4,8 +4,7 @@ export const TRANSACTIONS_ROUTE = "/transactions";
 export const BALANCE_ROUTE = "/balance";
 export const CONFIRM_TRANSACTION_ROUTE = "/confirm-transaction";
 export const ADDRESSES_ROUTE = "/addresses";
-export const REGISTER_IOVNAME_ROUTE = "/register-iovname";
-export const REGISTER_STARNAME_ROUTE = "/register-starname";
-export const REGISTER_NAME_ROUTE = "/register-name";
+export const REGISTER_PERSONALIZED_ADDRESS_ROUTE = "/register-personalized-address";
+export const REGISTER_STARNAME = "/register-starname";
 export const TERMS_ROUTE = "/terms";
 export const POLICY_ROUTE = "/policy";
