+++ packages/bierzo-wallet/src/routes/payment/index.tsx
@@ -9,7 +9,7 @@ import { generateSendTxRequest } from "../../communication/requestgenerators";
 import LedgerBillboardMessage from "../../components/BillboardMessage/LedgerBillboardMessage";
 import NeumaBillboardMessage from "../../components/BillboardMessage/NeumaBillboardMessage";
 import PageMenu from "../../components/PageMenu";
-import { isIovname, lookupRecipientAddressByName } from "../../logic/account";
+import { isIov, lookupRecipientAddressByName } from "../../logic/account";
 import { getCodecForChainId } from "../../logic/codec";
 import { RootState } from "../../store/reducers";
 import { BALANCE_ROUTE, TRANSACTIONS_ROUTE } from "../paths";
@@ -59,7 +59,7 @@ const Payment = (): JSX.Element => {
     const chainId = token.chainId;
 
     let recipient: Address;
-    if (isIovname(formValues[ADDRESS_FIELD])) {
+    if (isIov(formValues[ADDRESS_FIELD])) {
       const lookupResult = await lookupRecipientAddressByName(formValues[ADDRESS_FIELD], chainId);
 
       if (lookupResult === "name_not_found") {
