+++ packages/bierzo-wallet/src/routes/registerName/components/SelectAddressesTable.tsx
