+++ packages/sanes-browser-extension/src/extension/background/model/persona/index.ts
@@ -5,16 +5,10 @@ import {
   BnsUsernameNft,
   CreateProposalTx,
   isCreateProposalTx,
-  isRegisterAccountTx,
-  isRegisterDomainTx,
   isRegisterUsernameTx,
-  isReplaceAccountTargetsTx,
   isUpdateTargetsOfUsernameTx,
   isVoteTx,
-  RegisterAccountTx,
-  RegisterDomainTx,
   RegisterUsernameTx,
-  ReplaceAccountTargetsTx,
   UpdateTargetsOfUsernameTx,
   VoteTx,
 } from "@iov/bns";
@@ -48,9 +42,6 @@ export type SupportedTransaction =
   | SendTransaction
   | RegisterUsernameTx
   | UpdateTargetsOfUsernameTx
-  | RegisterDomainTx
-  | RegisterAccountTx
-  | ReplaceAccountTargetsTx
   | CreateProposalTx
   | VoteTx;
 
@@ -59,9 +50,6 @@ export function isSupportedTransaction(tx: UnsignedTransaction): tx is Supported
     isSendTransaction(tx) ||
     isRegisterUsernameTx(tx) ||
     isUpdateTargetsOfUsernameTx(tx) ||
-    isRegisterDomainTx(tx) ||
-    isRegisterAccountTx(tx) ||
-    isReplaceAccountTargetsTx(tx) ||
     isCreateProposalTx(tx) ||
     isVoteTx(tx)
   );
