+++ scripts/bnsd/genesis_app_state.json
@@ -198,14 +198,6 @@
       "owner": "cond:gov/rule/0000000000000002",
       "valid_username_name": "^[a-z0-9\\-_.]{3,64}$",
       "valid_username_label": "^iov$"
-    },
-    "account": {
-      "valid_name": "^[a-z0-9\\-_.]{3,64}$",
-      "valid_domain": "^[a-z0-9\\-_]{3,16}$",
-      "valid_blockchain_id": "^[a-z0-9A-Z-]+$",
-      "valid_blockchain_address": "^[a-z0-9A-Z]+$",
-      "domain_renew": "2h",
-      "owner": "cond:gov/rule/000000000000000002"
     }
   },
   "msgfee": [
@@ -276,31 +268,11 @@
       "ver": 1
     },
     {
-      "pkg": "account",
-      "ver": 1
-    },
-    {
       "pkg": "utils",
       "ver": 1
     },
     {
       "pkg": "validators",
-      "ver": 1
-    },
-    {
-      "pkg": "datamigration",
-      "ver": 1
-    },
-    {
-      "pkg": "preregistration",
-      "ver": 1
-    },
-    {
-      "pkg": "qualityscore",
-      "ver": 1
-    },
-    {
-      "pkg": "termdeposit",
       "ver": 1
     }
   ]
